Minor test standardizations and add dirs
