+++ backend/content/tests/discussion/test_discussion_create.py
@@ -8,7 +8,7 @@ from content.factories import DiscussionFactory
 pytestmark = pytest.mark.django_db
 
 
-def test_discussion_update():
+def test_discussion_create():
     client = APIClient()
     test_user = "test_user"
     test_pass = "test_pass"
@@ -17,20 +17,24 @@ def test_discussion_update():
     user.verified = True
     user.save()
 
-    thread = DiscussionFactory(created_by=user)
+    discussion_thread = DiscussionFactory(
+        title="Unique Title", category="Unique Category"
+    )
 
-    login = client.post(
+    # Login to get token.
+    login_response = client.post(
         path="/v1/auth/sign_in/", data={"username": test_user, "password": test_pass}
     )
 
-    assert login.status_code == 200
-    login_body = login.json()
+    assert login_response.status_code == 200
+
+    login_body = login_response.json()
     token = login_body["token"]
 
     client.credentials(HTTP_AUTHORIZATION=f"Token {token}")
-    response = client.put(
-        path=f"/v1/content/discussions/{thread.id}/",
-        data={"title": thread.title},
+    response = client.post(
+        path="/v1/content/discussions/",
+        data={"title": discussion_thread.title, "category": discussion_thread.category},
     )
 
-    assert response.status_code == 200
+    assert response.status_code == 201
