+++ backend/content/tests/discussion/test_discussion_partial_update.py
@@ -23,14 +23,15 @@ def test_discussion_partial_update():
         is_confirmed=True,
     )
 
-    # User login
-    login = client.post(
+    # Login to get token.
+    login_response = client.post(
         path="/v1/auth/sign_in/",
         data={"username": test_user, "password": test_pass},
     )
 
-    assert login.status_code == 200
-    login_body = login.json()
+    assert login_response.status_code == 200
+
+    login_body = login_response.json()
     token = login_body["token"]
 
     discussion_thread = DiscussionFactory(created_by=user)
@@ -53,5 +54,6 @@ def test_discussion_partial_update():
     )
 
     assert response.status_code == 403
+
     body = response.json()
     assert body["error"] == "You are not allowed to update this discussion."
