+++ backend/content/tests/resource/test_resource_partial_update.py
@@ -8,9 +8,9 @@ from content.factories import EntityLocationFactory, ResourceFactory
 pytestmark = pytest.mark.django_db
 
 
-def test_resource_update():
+def test_resource_partial_update():
     """
-    Test to update the resources.
+    Test to partially update resources.
     """
     client = APIClient()
 
@@ -27,35 +27,36 @@ def test_resource_update():
     resource = ResourceFactory(created_by=user)
     location = EntityLocationFactory()
 
-    login = client.post(
+    # Login to get token.
+    login_response = client.post(
         path="/v1/auth/sign_in/",
         data={"username": test_user, "password": test_pass},
     )
 
-    assert login.status_code == 200
-    login_body = login.json()
+    assert login_response.status_code == 200
+
+    login_body = login_response.json()
     token = login_body["token"]
 
     payload = {
         "name": "new_resource",
-        "description": "New Description",
-        "url": "https://activist.org/",
         "isPrivate": True,
         "termsChecked": True,
         "createdBy": user.id,
         "location": location.id,
     }
 
-    # Authorized owner tries to update the resources.
+    # Authorized owner tries to update the resource.
     client.credentials(HTTP_AUTHORIZATION=f"Token {token}")
-    response = client.put(path=f"/v1/content/resources/{resource.id}/", data=payload)
+    response = client.patch(path=f"/v1/content/resources/{resource.id}/", data=payload)
 
     assert response.status_code == 200
 
-    # Authorized non-owner tries to update the resources.
-    error_response = client.put(
+    # Authorized non-owner tries to update the resource.
+    error_response = client.patch(
         path=f"/v1/content/resources/{unowned_resource.id}/", data=payload
     )
     assert error_response.status_code == 403
+
     error_body = error_response.json()
     assert error_body["error"] == "You are not allowed to update this resource."
