+++ backend/content/tests/resource/test_resource_retrieve.py
@@ -3,42 +3,37 @@ import pytest
 from rest_framework.test import APIClient
 
 from authentication.factories import UserFactory
-from content.factories import DiscussionFactory
+from content.factories import ResourceFactory
 
 pytestmark = pytest.mark.django_db
 
 
-def test_disc_entry_create():
-    """
-    Test to check the creation of discussion entry.
-    """
+def test_resource_retrieve():
     client = APIClient()
 
     test_user = "test_user"
     test_pass = "test_pass"
-    user = UserFactory(
+    UserFactory(
         username=test_user,
         plaintext_password=test_pass,
         is_confirmed=True,
         verified=True,
     )
 
-    discussion_thread = DiscussionFactory(created_by=user)
+    resource = ResourceFactory(is_private=False)
 
-    # User login
-    login = client.post(
+    # Login to get token.
+    login_response = client.post(
         path="/v1/auth/sign_in/", data={"username": test_user, "password": test_pass}
     )
 
-    assert login.status_code == 200
-    login_body = login.json()
+    assert login_response.status_code == 200
+
+    login_body = login_response.json()
     token = login_body["token"]
 
     # Passing authorization header.
     client.credentials(HTTP_AUTHORIZATION=f"Token {token}")
-    response = client.post(
-        path="/v1/content/discussion_entries/",
-        data={"discussion": discussion_thread, "createdBy": user},
-    )
+    response = client.get(path=f"/v1/content/resources/{resource.id}/")
 
-    assert response.status_code == 201
+    assert response.status_code == 200
